requirements.txt
nudenet
aiosqlite
aiogram
async-lru
pytest
ImageHash
Pillow